@@ -1,6 +1,7 @@
-const { series, src, dest } = require("gulp");
-const svgSprite = require("gulp-svg-sprite");
-const rimraf = require("rimraf");
+import gulp from "gulp";
+import svgSprite from "gulp-svg-sprite";
+import rimraf from "rimraf";
+const { series, src, dest } = gulp;
 
 const spritesConfig = {
   shape: {
@@ -16,6 +17,8 @@ const spritesConfig = {
       padding: 0,
       box: "icon",
     },
+    meta: "./src/a11y-meta.yml",
+    dest: "../src/svgs/",
   },
   svg: {
     xmlDeclaration: false,
@@ -60,4 +63,5 @@ function buildSvgSprites() {
     .pipe(dest("sprites"));
 }
 
-exports.build = series(clean, buildSvgSprites);
+const build = series(clean, buildSvgSprites);
+export { clean, build };
