@@ -1,183 +1,243 @@
 import { IMXIcon } from "./common/icon";
+import aggregationIcon from "./svgs/aggregation.svg";
+import candidateResultsIcon from "./svgs/candidate-results.svg";
+import debateIcon from "./svgs/debate.svg";
+import electoralVotesIcon from "./svgs/electoral-votes.svg";
+import euParliamentIcon from "./svgs/eu-parliament.svg";
+import femaleCandidateIcon from "./svgs/female-candidate.svg";
+import hornIcon from "./svgs/horn.svg";
+import maleCandidateIcon from "./svgs/male-candidate.svg";
+import parliament1Icon from "./svgs/parliament-1.svg";
+import parliament2Icon from "./svgs/parliament-2.svg";
+import popularVotesIcon from "./svgs/popular-votes.svg";
+import protesterSignsIcon from "./svgs/protester-signs.svg";
+import resultsByCandidatesIcon from "./svgs/results-by-candidates.svg";
+import resultsPiechartIcon from "./svgs/results-piechart.svg";
+import scalesBalancedIcon from "./svgs/scales-balanced.svg";
+import scalesUnbalancedIcon from "./svgs/scales-unbalanced.svg";
+import statsIcon from "./svgs/stats.svg";
+import timelineIcon from "./svgs/timeline.svg";
+import useCaIcon from "./svgs/usa-ca.svg";
+import usaFlIcon from "./svgs/usa-fl.svg";
+import usaLower48StatesIcon from "./svgs/usa-lower-48-states.svg";
+import usaNyIcon from "./svgs/usa-ny.svg";
+import usaPartyDemocraticIcon from "./svgs/usa-party-democratic.svg";
+import usaPartyGreenIcon from "./svgs/usa-party-green.svg";
+import usaPartyLibertarianIcon from "./svgs/usa-party-libertarian.svg";
+import usaPartyRepublicanIcon from "./svgs/usa-party-republican.svg";
+import usaTxIcon from "./svgs/usa-tx.svg";
+import voteIcon from "./svgs/vote.svg";
+import whiteHouse1Icon from "./svgs/white-house-1.svg";
+import whiteHouse2Icon from "./svgs/white-house-2.svg";
 
 export const aggregation: IMXIcon = {
   name: "aggregation",
-  value: require("./svgs/aggregation.svg"),
+  value: aggregationIcon,
   categories: ["election"],
+  keywords: ["aggregation", "election results", "results", "elections", "merge", "converge"],
 };
 
 export const candidateResults: IMXIcon = {
     name: "candidate-results",
-    value: require("./svgs/candidate-results.svg"),
+    value: candidateResultsIcon,
     categories: ["election"],
+    keywords: ["front runner", "slate", "candidate", "results", "elections"],
 };
 
 export const debate: IMXIcon = {
     name: "debate",
-    value: require("./svgs/debate.svg"),
+    value: debateIcon,
     categories: ["election"],
+    keywords: ["discussion", "dispute", "argument", "elections", "battle", "fight", "boxing"],
 };
 
 export const electoralVotes: IMXIcon = {
     name: "electoral-votes",
-    value: require("./svgs/electoral-votes.svg"),
+    value: electoralVotesIcon,
     categories: ["election"],
+    keywords: ["electoral college", "electoral", "votes", "presidential election", "elections"],
 };
 
 export const euParliament: IMXIcon = {
     name: "eu-parliament",
-    value: require("./svgs/eu-parliament.svg"),
+    value: euParliamentIcon,
     categories: ["election"],
+    keywords: ["European Union", "Europe", "parliament", "EU", "government", "politics"],
 };
 
 export const femaleCandidate: IMXIcon = {
     name: "female-candidate",
-    value: require("./svgs/female-candidate.svg"),
+    value: femaleCandidateIcon,
     categories: ["election"],
+    keywords: ["female", "woman", "elections", "candidate", "speaker", "presenter"],
 };
 
 export const horn: IMXIcon = {
     name: "horn",
-    value: require("./svgs/horn.svg"),
+    value: hornIcon,
     categories: ["election"],
+    keywords: ["broadcast", "megaphone", "speaker", "volume", "elections"],
 };
 
 export const maleCandidate: IMXIcon = {
     name: "male-candidate",
-    value: require("./svgs/male-candidate.svg"),
+    value: maleCandidateIcon,
     categories: ["election"],
+    keywords: ["male", "man", "elections", "candidate", "speaker", "presenter"],
 };
 
 export const parliament1: IMXIcon = {
     name: "parliament-1",
-    value: require("./svgs/parliament-1.svg"),
+    value: parliament1Icon,
     categories: ["election"],
+    keywords: ["parliament", "capitol", "goverment", "house", "politics", "Europe", "temple"],
 };
 
 export const parliament2: IMXIcon = {
     name: "parliament-2",
-    value: require("./svgs/parliament-2.svg"),
+    value: parliament2Icon,
     categories: ["election"],
+    keywords: ["parliament", "capitol", "goverment", "house", "politics", "Berlin", "gate", "monument"],
 };
 
 export const popularVotes: IMXIcon = {
     name: "popular-votes",
-    value: require("./svgs/popular-votes.svg"),
+    value: popularVotesIcon,
     categories: ["election"],
+    keywords: ["vote", "ballot", "cast", "station", "election"],
 }
 
 export const protesterSigns: IMXIcon = {
     name: "protester-signs",
-    value: require("./svgs/protester-signs.svg"),
+    value: protesterSignsIcon,
     categories: ["election"],
+    keywords: ["protest", "protester", "demonstration", "signs", "revolt", "turmoil", "riot"],
 }
 
 export const resultsByCandidates: IMXIcon = {
     name: "results-by-candidates",
-    value: require("./svgs/results-by-candidates.svg"),
+    value: resultsByCandidatesIcon,
     categories: ["election"],
+    keywords: ["results", "chart", "pie"],
 };
 
 export const resultsPiechart: IMXIcon = {
     name: "results-piechart",
-    value: require("./svgs/results-piechart.svg"),
+    value: resultsPiechartIcon,
     categories: ["election"],
+    keywords: ["results", "chart", "pie"],
 };
 
 export const scalesBalanced: IMXIcon = {
     name: "scales-balanced",
-    value: require("./svgs/scales-balanced.svg"),
+    value: scalesBalancedIcon,
     categories: ["election"],
+    keywords: ["scales", "justice", "balance", "legal", "weight"],
 };
 
 export const scalesUnbalanced: IMXIcon = {
     name: "scales-unbalanced",
-    value: require("./svgs/scales-unbalanced.svg"),
+    value: scalesUnbalancedIcon,
     categories: ["election"],
+    keywords: ["scales", "unfair", "unbalanced", "off-balance", "legal", "weight"],
 };
 
 export const stats: IMXIcon = {
     name: "stats",
-    value: require("./svgs/stats.svg"),
+    value: statsIcon,
     categories: ["election"],
+    keywords: ["statistics", "chart", "screen", "display", "monitor"],
 };
 
 export const timeline: IMXIcon = {
     name: "timeline",
-    value: require("./svgs/timeline.svg"),
+    value: timelineIcon,
     categories: ["election"],
+    keywords: ["timeline", "events"],
 };
 
 export const usaCa: IMXIcon = {
     name: "usa-ca",
-    value: require("./svgs/usa-ca.svg"),
+    value: useCaIcon,
     categories: ["election"],
+    keywords: ["USA", "state", "California", "ca"],
 };
 
 export const usaFl: IMXIcon = {
     name: "usa-fl",
-    value: require("./svgs/usa-fl.svg"),
+    value: usaFlIcon,
     categories: ["election"],
+    keywords: ["USA", "state", "Florida", "fl"],
 };
 
 export const usaLower48States: IMXIcon = {
     name: "usa-lower-48-states",
-    value: require("./svgs/usa-lower-48-states.svg"),
+    value: usaLower48StatesIcon,
     categories: ["election"],
+    keywords: ["USA", "state", "lower", "48"],
 };
 
 export const usaNy: IMXIcon = {
     name: "usa-ny",
-    value: require("./svgs/usa-ny.svg"),
+    value: usaNyIcon,
     categories: ["election"],
+    keywords: ["USA", "state", "New York", "ny"],
 };
 
 export const usaPartyDemocratic: IMXIcon = {
     name: "usa-party-democratic",
-    value: require("./svgs/usa-party-democratic.svg"),
+    value: usaPartyDemocraticIcon,
     categories: ["election"],
+    keywords: ["USA", "party", "democratic", "democrat", "democrats", "donkey", "politics"],
 };
 
 export const usaPartyGreen: IMXIcon = {
     name: "usa-party-green",
-    value: require("./svgs/usa-party-green.svg"),
+    value: usaPartyGreenIcon,
     categories: ["election"],
+    keywords: ["USA", "party", "green", "politics"],
 };
 
 export const usaPartyLibertarian: IMXIcon = {
     name: "usa-party-libertarian",
-    value: require("./svgs/usa-party-libertarian.svg"),
+    value: usaPartyLibertarianIcon,
     categories: ["election"],
+    keywords: ["USA", "party", "libertarian", "hedgehog", "politics"],
 };
 
 export const usaPartyRepublican: IMXIcon = {
     name: "usa-party-republican",
-    value: require("./svgs/usa-party-republican.svg"),
+    value: usaPartyRepublicanIcon,
     categories: ["election"],
+    keywords: ["USA", "party", "republican", "republicans", "elephant", "politics"],
 };
 
 export const usaTx: IMXIcon = {
     name: "usa-tx",
-    value: require("./svgs/usa-tx.svg"),
+    value: usaTxIcon,
     categories: ["election"],
+    keywords: ["USA", "state", "Texas", "tx"],
 };
 
 export const vote: IMXIcon = {
     name: "vote",
-    value: require("./svgs/vote.svg"),
+    value: voteIcon,
     categories: ["election"],
+    keywords: ["vote", "cast", "ballot", "box", "station", "election"],
 };
 
 export const whiteHouse1: IMXIcon = {
     name: "white-house-1",
-    value: require("./svgs/white-house-1.svg"),
+    value: whiteHouse1Icon,
     categories: ["election"],
+    keywords: ["white house", "president", "house", "government", "Washington", "DC"],
 };
 
 export const whiteHouse2: IMXIcon = {
     name: "white-house-2",
-    value: require("./svgs/white-house-2.svg"),
+    value: whiteHouse2Icon,
     categories: ["election"],
+    keywords: ["white house", "president", "house", "government", "Washington", "DC", "Athens", "acropolis"],
 };
 
 export const elections = [
