@@ -1,321 +1,427 @@
 import { IMXIcon } from "./common/icon";
+import atmDollarIcon from "./svgs/atm-dollar.svg";
+import atmEuroIcon from "./svgs/atm-euro.svg";
+import atmPoundIcon from "./svgs/atm-pound.svg";
+import atmIcon from "./svgs/atm.svg";
+import atmYenIcon from "./svgs/atm-yen.svg";
+import auctionIcon from "./svgs/auction.svg";
+import awardIcon from "./svgs/award.svg";
+import bankSafeIcon from "./svgs/bank-safe.svg";
+import barcodeIcon from "./svgs/barcode.svg";
+import bearMarketIcon from "./svgs/bear-market.svg";
+import billPaidIcon from "./svgs/bill-paid.svg";
+import budgetSpendingIcon from "./svgs/budget-spending.svg";
+import bullMarketIcon from "./svgs/bull-market.svg";
+import businessDevelopmentIcon from "./svgs/business-development.svg";
+import calculatorIcon from "./svgs/calculator.svg";
+import cashIcon from "./svgs/cash-1.svg";
+import cashAltIcon from "./svgs/cash-2.svg";
+import cashOnlyIcon from "./svgs/cash-only.svg";
+import certificateIcon from "./svgs/certificate.svg";
+import commoditiesIcon from "./svgs/commodities.svg";
+import creditCardsIcon from "./svgs/credit-cards.svg";
+import creditControlIcon from "./svgs/credit-control.svg";
+import crowdfundingIcon from "./svgs/crowdfunding.svg";
+import debtIcon from "./svgs/debt.svg";
+import deliveryIcon from "./svgs/delivery.svg";
+import diamondCircledIcon from "./svgs/diamond-circled.svg";
+import diamondIcon from "./svgs/diamond.svg";
+import discountIcon from "./svgs/discount.svg";
+import dividentsIcon from "./svgs/dividents.svg";
+import dollarCircledIcon from "./svgs/dollar-circled.svg";
+import dropShippingIcon from "./svgs/drop-shipping.svg";
+import euroCircledIcon from "./svgs/euro-circled.svg";
+import faceIdIcon from "./svgs/face-id.svg";
+import financialInsuranceIcon from "./svgs/financial-insurance.svg";
+import foreignExchangeIcon from "./svgs/foreign-exchange.svg";
+import giftIcon from "./svgs/gift.svg";
+import goldIcon from "./svgs/gold.svg";
+import ideaGenerationIcon from "./svgs/idea-generation.svg";
+import invoiceIcon from "./svgs/invoice.svg";
+import legalDocumentIcon from "./svgs/legal-document.svg";
+import loanIcon from "./svgs/loan.svg";
+import moneyBagIcon from "./svgs/money-bag.svg";
+import oilIcon from "./svgs/oil.svg";
+import payDateIcon from "./svgs/pay-date.svg";
+import piggyBankIcon from "./svgs/piggy-bank.svg";
+import poundCircledIcon from "./svgs/pound-circled.svg";
+import projectRevenueIcon from "./svgs/project-revenue.svg";
+import saleIcon from "./svgs/sale.svg";
+import securityIcon from "./svgs/security-1.svg";
+import securityAltIcon from "./svgs/security-2.svg";
+import strategyIcon from "./svgs/strategy.svg";
+import wireTransferIcon from "./svgs/wire-transfer.svg";
+import yenCircledIcon from "./svgs/yen-circled.svg";
 
 export const atmDollar: IMXIcon = {
   name: "atm-dollar",
-  value: require("./svgs/atm-dollar.svg"),
+  value: atmDollarIcon,
   categories: ["finance"],
+  keywords:["atm", "usd", "dollar", "bank", "cash", "money", "payment", "deposit", "withdraw", "machine"],
 };
 
 export const atmEuro: IMXIcon = {
   name: "atm-euro",
-  value: require("./svgs/atm-euro.svg"),
+  value: atmEuroIcon,
   categories: ["finance"],
+  keywords:["atm", "eur", "euro", "bank", "cash", "money", "payment", "deposit", "withdraw", "machine"],
 };
 
 export const atmPound: IMXIcon = {
   name: "atm-pound",
-  value: require("./svgs/atm-pound.svg"),
+  value: atmPoundIcon,
   categories: ["finance"],
+  keywords:["atm", "gbp", "pound", "sterling", "bank", "cash", "money", "payment", "deposit", "withdraw", "machine"],
 };
 
 export const atm: IMXIcon = {
   name: "atm",
-  value: require("./svgs/atm.svg"),
+  value: atmIcon,
   categories: ["finance"],
+  keywords:["atm", "bank", "cash", "money", "payment", "deposit", "withdraw", "machine"],
 };
 
 export const atmYen: IMXIcon = {
   name: "atm-yen",
-  value: require("./svgs/atm-yen.svg"),
+  value: atmYenIcon,
   categories: ["finance"],
+  keywords:["atm", "jpy", "yen", "bank", "cash", "money", "payment", "deposit", "withdraw", "machine"],
 };
 
 export const auction: IMXIcon = {
   name: "auction",
-  value: require("./svgs/auction.svg"),
+  value: auctionIcon,
   categories: ["finance"],
+  keywords:["auction", "sale", "gavel", "law", "judge", "hammer"],
 };
 
 export const award: IMXIcon = {
   name: "award",
-  value: require("./svgs/award.svg"),
+  value: awardIcon,
   categories: ["finance"],
+  keywords:["award", "medal", "winner", "first", "gold", "sport"],
 };
 
 export const bankSafe: IMXIcon = {
   name: "bank-safe",
-  value: require("./svgs/bank-safe.svg"),
+  value: bankSafeIcon,
   categories: ["finance"],
+  keywords:["bank", "safe", "savings", "secure", "vault", "deposit",  "locker", "safe box"],
 };
 
 export const barcode: IMXIcon = {
   name: "barcode",
-  value: require("./svgs/barcode.svg"),
+  value: barcodeIcon,
   categories: ["finance"],
+  keywords:["barcode", "code", "scan", "marketing", "commerce"],
 };
 
 export const bearMarket: IMXIcon = {
   name: "bear-market",
-  value: require("./svgs/bear-market.svg"),
+  value: bearMarketIcon,
   categories: ["finance"],
+  keywords:["bear", "market", "decreasing", "financial", "economy", "animal", "stock"],
 };
 
 export const billPaid: IMXIcon = {
   name: "bill-paid",
-  value: require("./svgs/bill-paid.svg"),
+  value: billPaidIcon,
   categories: ["finance"],
+  keywords:["bill", "paid", "check", "invoice", "payment", "purchase", "accounting", "billing"],
 };
 
 export const budgetSpending: IMXIcon = {
   name: "budget-spending",
-  value: require("./svgs/budget-spending.svg"),
+  value: budgetSpendingIcon,
   categories: ["finance"],
+  keywords:["budget", "spending", "usd", "dollar", "accounting", "expense", "investment", "management", "ballance"],
 };
 
 export const bullMarket: IMXIcon = {
   name: "bull-market",
-  value: require("./svgs/bull-market.svg"),
+  value: bullMarketIcon,
   categories: ["finance"],
+  keywords:["bull", "market", "increasing", "financial", "economy", "animal", "stock"],
 };
 
 export const businessDevelopment: IMXIcon = {
   name: "business-development",
-  value: require("./svgs/business-development.svg"),
+  value: businessDevelopmentIcon,
   categories: ["finance"],
+  keywords:["business", "development", "global", "worldwide", "solutions", "organisation", "innovation", "management"],
 };
 
 export const calculator: IMXIcon = {
   name: "calculator",
-  value: require("./svgs/calculator.svg"),
+  value: calculatorIcon,
   categories: ["finance"],
+  keywords:["calculator", "math", "school", "business", "accounting"],
 };
 
 export const cash: IMXIcon = {
   name: "cash",
-  value: require("./svgs/cash-1.svg"),
+  value: cashIcon,
   categories: ["finance"],
+  keywords:["cash", "money", "notes", "bank", "pay", "payment", "banknote", "currency"],
 };
 
 export const cashAlt: IMXIcon = {
   name: "cash-alt",
-  value: require("./svgs/cash-2.svg"),
+  value: cashAltIcon,
   categories: ["finance"],
+  keywords:["cash", "money", "notes", "bank", "pay", "payment", "banknote", "currency", "usd", "dollar"],
 };
 
 export const cashOnly: IMXIcon = {
   name: "cash-only",
-  value: require("./svgs/cash-only.svg"),
+  value: cashOnlyIcon,
   categories: ["finance"],
+  keywords:["cash", "no card", "payment", "money", "pay", "cash only"],
 };
 
 export const certificate: IMXIcon = {
   name: "certificate",
-  value: require("./svgs/certificate.svg"),
+  value: certificateIcon,
   categories: ["finance"],
+  keywords:["certificate", "medal", "badge", "seal", "prize", "award", "guarantee", "quality", "stamp", "warranty", "collection"],
 };
 
 export const commodities: IMXIcon = {
   name: "commodities",
-  value: require("./svgs/commodities.svg"),
+  value: commoditiesIcon,
   categories: ["finance"],
+  keywords:["commodities", "oil", "gold", "trade", "commodity", "crude", "market", "price", "invest"],
 };
 
 export const creditCards: IMXIcon = {
   name: "credit-cards",
-  value: require("./svgs/credit-cards.svg"),
+  value: creditCardsIcon,
   categories: ["finance"],
+  keywords:["credit", "debit", "cards", "payment", "shopping", "business", "money", "cash"],
 };
 
 export const creditControl: IMXIcon = {
   name: "credit-control",
-  value: require("./svgs/credit-control.svg"),
+  value: creditControlIcon,
   categories: ["finance"],
+  keywords:["credit", "control", "watch", "eye"],
 };
 
 export const crowdfunding: IMXIcon = {
   name: "crowdfunding",
-  value: require("./svgs/crowdfunding.svg"),
+  value: crowdfundingIcon,
   categories: ["finance"],
+  keywords:["crowdfunding", "mutual", "fund", "contribution", "donation", "money", "usd", "dollar", "revenue", "budget", "fundraising"],
 };
 
 export const debt: IMXIcon = {
   name: "debt",
-  value: require("./svgs/debt.svg"),
+  value: debtIcon,
   categories: ["finance"],
+  keywords:["debt", "financial", "usd", "dollar", "money", "request"],
 };
 
 export const delivery: IMXIcon = {
   name: "delivery",
-  value: require("./svgs/delivery.svg"),
+  value: deliveryIcon,
   categories: ["finance"],
+  keywords:["delivery", "truck", "logistics", "van", "courier", "shipping", "service", "vehicle"],
 };
 
 export const diamondCircled: IMXIcon = {
   name: "diamond-circled",
-  value: require("./svgs/diamond-circled.svg"),
+  value: diamondCircledIcon,
   categories: ["finance"],
+  keywords:["diamond", "gem", "precious", "jewellery", "circled", "jewel", "gemstone"],
 };
 
 export const diamond: IMXIcon = {
   name: "diamond",
-  value: require("./svgs/diamond.svg"),
+  value: diamondIcon,
   categories: ["finance"],
+  keywords:["diamond", "gem", "precious", "jewellery", "jewel", "gemstone"],
 };
 
 export const discount: IMXIcon = {
   name: "discount",
-  value: require("./svgs/discount.svg"),
+  value: discountIcon,
   categories: ["finance"],
+  keywords:["discount", "percent", "percentage", "sale", "badge", "off", "%", "rebate", "subsidy"],
 };
 
 export const dividents: IMXIcon = {
   name: "dividents",
-  value: require("./svgs/dividents.svg"),
+  value: dividentsIcon,
   categories: ["finance"],
+  keywords:["dividents", "%", "percent", "percentage", "income", "rate", "cut", "intrest", "investment"],
 };
 
 export const dollarCircled: IMXIcon = {
   name: "dollar-circled",
-  value: require("./svgs/dollar-circled.svg"),
+  value: dollarCircledIcon,
   categories: ["finance"],
+  keywords:["usd", "dollar", "circled", "currency"],
 };
 
 export const dropShipping: IMXIcon = {
   name: "drop-shipping",
-  value: require("./svgs/drop-shipping.svg"),
+  value: dropShippingIcon,
   categories: ["finance"],
+  keywords:["drop", "shipping", "dropshipping", "box", "parachute", "business", "logistics", "crate", "air", "supply", "airdrop", "fortnite", "pubg", "battle royale", "delivery", "package"],
 };
 
 export const euroCircled: IMXIcon = {
   name: "euro-circled",
-  value: require("./svgs/euro-circled.svg"),
+  value: euroCircledIcon,
   categories: ["finance"],
+  keywords:["eur", "euro", "circled", "currency", "eu"],
 };
 
 export const faceId: IMXIcon = {
   name: "face-id",
-  value: require("./svgs/face-id.svg"),
+  value: faceIdIcon,
   categories: ["finance"],
+  keywords:["faceid", "biometric", "identity", "security", "face"],
 };
 
 export const financialInsurance: IMXIcon = {
   name: "financial-insurance",
-  value: require("./svgs/financial-insurance.svg"),
+  value: financialInsuranceIcon,
   categories: ["finance"],
+  keywords:["financial", "insurance", "business", "money", "protection", "security"],
 };
 
 export const foreignExchange: IMXIcon = {
   name: "foreign-exchange",
-  value: require("./svgs/foreign-exchange.svg"),
+  value: foreignExchangeIcon,
   categories: ["finance"],
+  keywords:["foreign", "exchange", "currency", "forex", "swap", "trade", "bank", "dollar", "usd", "euro", "eur", "money", "rate"],
 };
 
 export const gift: IMXIcon = {
   name: "gift",
-  value: require("./svgs/gift.svg"),
+  value: giftIcon,
   categories: ["finance"],
+  keywords:["gift", "present", "box", "christmas", "bonus", "give"],
 };
 
 export const gold: IMXIcon = {
   name: "gold",
-  value: require("./svgs/gold.svg"),
+  value: goldIcon,
   categories: ["finance"],
+  keywords:[""],
 };
 
 export const ideaGeneration: IMXIcon = {
   name: "idea-generation",
-  value: require("./svgs/idea-generation.svg"),
+  value: ideaGenerationIcon,
   categories: ["finance"],
+  keywords:["gold", "business", "currency", "finance", "gold bars", "treasure", "gold blocks", "gold cubes", "gold bricks", "gold ingots", "gold biscuits"],
 };
 
 export const invoice: IMXIcon = {
   name: "invoice",
-  value: require("./svgs/invoice.svg"),
+  value: invoiceIcon,
   categories: ["finance"],
+  keywords:["invoice", "accounting", "bill", "business", "document", "file", "billing", "payment", "purchase"],
 };
 
 export const legalDocument: IMXIcon = {
   name: "legal-document",
-  value: require("./svgs/legal-document.svg"),
+  value: legalDocumentIcon,
   categories: ["finance"],
+  keywords:["legal", "document", "file", "contract", "paper", "law", "court"],
 };
 
 export const loan: IMXIcon = {
   name: "loan",
-  value: require("./svgs/loan.svg"),
+  value: loanIcon,
   categories: ["finance"],
+  keywords:["loan", "collateral", "convinience loan", "mortgage", "personal loan", "lending", "money", "dollar", "usd"],
 };
 
 export const moneyBag: IMXIcon = {
   name: "money-bag",
-  value: require("./svgs/money-bag.svg"),
+  value: moneyBagIcon,
   categories: ["finance"],
+  keywords:["money", "bag", "cash", "coin", "currency", "dollar", "usd", "investment"],
 };
 
 export const oil: IMXIcon = {
   name: "oil",
-  value: require("./svgs/oil.svg"),
+  value: oilIcon,
   categories: ["finance"],
+  keywords:["oil", "crude", "commodities", "barrel", "oil container", "petroleum"],
 };
 
 export const payDate: IMXIcon = {
   name: "pay-date",
-  value: require("./svgs/pay-date.svg"),
+  value: payDateIcon,
   categories: ["finance"],
+  keywords:["pay date", "calendar", "salary", "card", "payday", "payment", "finance", "schedule", "plan", "direct debit"],
 };
 
 export const piggyBank: IMXIcon = {
   name: "piggy-bank",
-  value: require("./svgs/piggy-bank.svg"),
+  value: piggyBankIcon,
   categories: ["finance"],
+  keywords:["piggy bank", "money", "save", "savings", "piggybank", "business", "cash", "pig", "coin", "coins", "bank", "deposit", "piglet"],
 };
 
 export const poundCircled: IMXIcon = {
   name: "pound-circled",
-  value: require("./svgs/pound-circled.svg"),
+  value: poundCircledIcon,
   categories: ["finance"],
+  keywords:["gbp", "pound", "sterling", "circled", "currency"],
 };
 
 export const projectRevenue: IMXIcon = {
   name: "project-revenue",
-  value: require("./svgs/project-revenue.svg"),
+  value: projectRevenueIcon,
   categories: ["finance"],
+  keywords:["project", "revenue", "turnover", "income", "dollar", "usd", "money", "capital", "earnings", "profit", "automation", "make money"],
 };
 
 export const sale: IMXIcon = {
   name: "sale",
-  value: require("./svgs/sale.svg"),
+  value: saleIcon,
   categories: ["finance"],
+  keywords:["sale", "circled", "rebate", "discount", "deal", "commerce", "shopping", "promo", "flash", "flashsale"],
 };
 
 export const security: IMXIcon = {
   name: "security",
-  value: require("./svgs/security-1.svg"),
+  value: securityIcon,
   categories: ["finance"],
+  keywords:["security", "briefcase", "case", "guard", "insurance", "protection", "shield", "bag", "luggage", "document", "save"],
 };
 
 export const securityAlt: IMXIcon = {
   name: "security-alt",
-  value: require("./svgs/security-2.svg"),
+  value: securityAltIcon,
   categories: ["finance"],
+  keywords:["security", "briefcase", "usd", "dollar", "case", "guard", "insurance", "protection", "bag", "luggage", "document", "save"],
 };
 
 export const strategy: IMXIcon = {
   name: "strategy",
-  value: require("./svgs/strategy.svg"),
+  value: strategyIcon,
   categories: ["finance"],
+  keywords:["strategy", "business", "corporate", "strategic", "management", "plan", "planning", "chess", "game", "play"],
 };
 
 export const wireTransfer: IMXIcon = {
   name: "wire-transfer",
-  value: require("./svgs/wire-transfer.svg"),
+  value: wireTransferIcon,
   categories: ["finance"],
+  keywords:["wire", "money", "transfer", "currency", "exchange", "euro", "eur", "dollar", "usd", "bank", "payment", "trade", "transaction"],
 };
 
 export const yenCircled: IMXIcon = {
   name: "yen-circled",
-  value: require("./svgs/yen-circled.svg"),
+  value: yenCircledIcon,
   categories: ["finance"],
+  keywords:["jpy", "yen", "circled", "currency"],
 };
 
 export const finance = [
