@@ -1,435 +1,579 @@
 import { IMXIcon } from "./common/icon";
+import accessibilityIcon from "./svgs/accessibility.svg";
+import aidsRibbonIcon from "./svgs/aids-ribbon.svg";
+import ambulanceIcon from "./svgs/ambulance.svg";
+import anxietyIcon from "./svgs/anxiety.svg";
+import babyBottleIcon from "./svgs/baby-bottle.svg";
+import bacteriaIcon from "./svgs/bacteria.svg";
+import bandageIcon from "./svgs/bandage.svg";
+import biohazardIcon from "./svgs/biohazard.svg";
+import bloodIcon from "./svgs/blood.svg";
+import brailleIcon from "./svgs/braille.svg";
+import breastfeedingIcon from "./svgs/breastfeeding.svg";
+import caduceusIcon from "./svgs/caduceus.svg";
+import cancerRibbonIcon from "./svgs/cancer-ribbon.svg";
+import clinicIcon from "./svgs/clinic.svg";
+import condomIcon from "./svgs/condom.svg";
+import condomUsedIcon from "./svgs/condom-used.svg";
+import contraceptionIcon from "./svgs/contraception.svg";
+import coronavirusIcon from "./svgs/coronavirus.svg";
+import deafnessIcon from "./svgs/deafness.svg";
+import dentalFlossIcon from "./svgs/dental-floss.svg";
+import dentalImplantIcon from "./svgs/dental-implant.svg";
+import diabetesIcon from "./svgs/diabetes.svg";
+import dnaIcon from "./svgs/dna.svg";
+import doctorsBagIcon from "./svgs/doctors-bag.svg";
+import doctorIcon from "./svgs/doctor.svg";
+import doseIcon from "./svgs/dose.svg";
+import embryoIcon from "./svgs/embryo.svg";
+import fertilizationIcon from "./svgs/fertilization.svg";
+import fitnessWatchIcon from "./svgs/fitness-watch.svg";
+import gasMaskIcon from "./svgs/gas-mask.svg";
+import hearingAidIcon from "./svgs/hearing-aid.svg";
+import heartCareIcon from "./svgs/heart-care.svg";
+import heartMonitorIcon from "./svgs/heart-monitor.svg";
+import hospitalBedIcon from "./svgs/hospital-bed.svg";
+import hospitalSignIcon from "./svgs/hospital-sign.svg";
+import hospitalIcon from "./svgs/hospital.svg";
+import keepAwayFromChildrenIcon from "./svgs/keep-away-from-children.svg";
+import maskIcon from "./svgs/mask-1.svg";
+import maskAltIcon from "./svgs/mask-2.svg";
+import messageIcon from "./svgs/massage.svg";
+import medicalBookIcon from "./svgs/medical-book.svg";
+import medicineIcon from "./svgs/medicine.svg";
+import microorganismsIcon from "./svgs/microorganisms.svg";
+import microscopeIcon from "./svgs/microscope.svg";
+import nurseIcon from "./svgs/nurse.svg";
+import pacifierIcon from "./svgs/pacifier.svg";
+import pillIcon from "./svgs/pill.svg";
+import plasticSurgeryIcon from "./svgs/plastic-surgery.svg";
+import poopIcon from "./svgs/poop.svg";
+import pregnancyNoAlcoholIcon from "./svgs/pregnancy-no-alcohol.svg";
+import pulseIcon from "./svgs/pulse.svg";
+import radioactiveIcon from "./svgs/radioactive.svg";
+import rhNegativeIcon from "./svgs/rh-negative.svg";
+import rhPositiveIcon from "./svgs/rh-positive.svg";
+import rodOfAsclepiusIcon from "./svgs/rod-of-asclepius.svg";
+import signLanguageIcon from "./svgs/sign-language.svg";
+import sneezeIcon from "./svgs/sneeze.svg";
+import stethoscopeIcon from "./svgs/stethoscope.svg";
+import stretcherIcon from "./svgs/stretcher.svg";
+import syringeIcon from "./svgs/syringe.svg";
+import teethIcon from "./svgs/teeth.svg";
+import thermometerIcon from "./svgs/thermometer.svg";
+import toiletPaperIcon from "./svgs/toilet-paper.svg";
+import toothCrackedIcon from "./svgs/tooth-cracked.svg";
+import toothIcon from "./svgs/tooth.svg";
+import veterinarianIcon from "./svgs/veterinarian.svg";
+import vialIcon from "./svgs/vial.svg";
+import virusProtectionIcon from "./svgs/virus-protection.svg";
+import virusStopIcon from "./svgs/virus-stop.svg";
+import virusIcon from "./svgs/virus.svg";
+import visuallyImpairedIcon from "./svgs/visually-impaired.svg";
+import weightScaleIcon from "./svgs/weight-scale.svg";
 
 export const accessibility: IMXIcon = {
   name: "accessibility",
-  value: require("./svgs/accessibility.svg"),
+  value: accessibilityIcon,
   categories: ["health"],
+  keywords:["accessibility", "wheelchair", "disabled", "handicapped", "chairbound", "impaired"],
 };
 
 export const aidsRibbon: IMXIcon = {
   name: "aids-ribbon",
-  value: require("./svgs/aids-ribbon.svg"),
+  value: aidsRibbonIcon,
   categories: ["health"],
+  keywords:["aids", "ribbon", "hiv", "care", "community"],
 };
 
 export const ambulance: IMXIcon = {
   name: "ambulance",
-  value: require("./svgs/ambulance.svg"),
+  value: ambulanceIcon,
   categories: ["health"],
+  keywords:["ambulance", "car", "medicine", "emergency", "vehicle", "van", "healthcare", "medical", "rescue"],
 };
 
 export const anxiety: IMXIcon = {
   name: "anxiety",
-  value: require("./svgs/anxiety.svg"),
+  value: anxietyIcon,
   categories: ["health"],
+  keywords:["anxiety", "brainstorm", "mental", "depression", "despair", "fatigue", "pressure", "stress", "tension", "worried"],
 };
 
 export const babyBottle: IMXIcon = {
   name: "baby-bottle",
-  value: require("./svgs/baby-bottle.svg"),
+  value: babyBottleIcon,
   categories: ["health"],
+  keywords:["baby", "bottle", "feeding", "milk"],
 };
 
 export const bacteria: IMXIcon = {
   name: "bacteria",
-  value: require("./svgs/bacteria.svg"),
+  value: bacteriaIcon,
   categories: ["health"],
+  keywords:["bacteria", "virus", "covid", "coronavirus", "microbe", "germs", "science", "biology"],
 };
 
 export const bandage: IMXIcon = {
   name: "bandage",
-  value: require("./svgs/bandage.svg"),
+  value: bandageIcon,
   categories: ["health"],
+  keywords:["bandage", "band", "bandaid", "plaster", "aid", "wound", "first aid", "healing"],
 };
 
 export const biohazard: IMXIcon = {
   name: "biohazard",
-  value: require("./svgs/biohazard.svg"),
+  value: biohazardIcon,
   categories: ["health"],
+  keywords:["biohazard", "waste"],
 };
 
 export const blood: IMXIcon = {
   name: "blood",
-  value: require("./svgs/blood.svg"),
+  value: bloodIcon,
   categories: ["health"],
+  keywords:["blood", "drop", "donation", "droplet", ],
 };
 
 export const braille: IMXIcon = {
   name: "braille",
-  value: require("./svgs/braille.svg"),
+  value: brailleIcon,
   categories: ["health"],
+  keywords:["braile", "alphabet", "blind", "read", "accessibility", "impaired"],
 };
 
 export const breastfeeding: IMXIcon = {
   name: "breastfeeding",
-  value: require("./svgs/breastfeeding.svg"),
+  value: breastfeedingIcon,
   categories: ["health"],
+  keywords:["breastfeeding", "mother", "nurse"],
 };
 
 export const caduceus: IMXIcon = {
   name: "caduceus",
-  value: require("./svgs/caduceus.svg"),
+  value: caduceusIcon,
   categories: ["health"],
+  keywords:["caduceus", "medical", "asclepiuss", "pharmacy", "doctor", "healthcare"],
 };
 
 export const cancerRibbon: IMXIcon = {
   name: "cancer-ribbon",
-  value: require("./svgs/cancer-ribbon.svg"),
+  value: cancerRibbonIcon,
   categories: ["health"],
+  keywords:["cancer", "ribbon", "awareness", "breast"],
 };
 
 export const clinic: IMXIcon = {
   name: "clinic",
-  value: require("./svgs/clinic.svg"),
+  value: clinicIcon,
   categories: ["health"],
+  keywords:["clinic", "hospital", "medical", "building", "therapy", "health center"],
 };
 
 export const condom: IMXIcon = {
   name: "condom",
-  value: require("./svgs/condom.svg"),
+  value: condomIcon,
   categories: ["health"],
+  keywords:["condom", "contraceptive", "sex", "hiv", "latex", "preservative", "healthcare", "sexual", "safe"],
 };
 
 export const condomUsed: IMXIcon = {
   name: "condom-used",
-  value: require("./svgs/condom-used.svg"),
+  value: condomUsedIcon,
   categories: ["health"],
+  keywords:["condom", "contraceptive", "sex", "hiv", "latex", "preservative", "healthcare", "sexual", "condom used", "semen", "sperm", "safe"],
 };
 
 export const contraception: IMXIcon = {
   name: "contraception",
-  value: require("./svgs/contraception.svg"),
+  value: contraceptionIcon,
   categories: ["health"],
+  keywords:[""],
 };
 
 export const coronavirus: IMXIcon = {
   name: "coronavirus",
-  value: require("./svgs/coronavirus.svg"),
+  value: coronavirusIcon,
   categories: ["health"],
+  keywords:["coronavirus", "virus", "covid", "pandemic", "corona", "germ", "infection"],
 };
 
 export const deafness: IMXIcon = {
   name: "deafness",
-  value: require("./svgs/deafness.svg"),
+  value: deafnessIcon,
   categories: ["health"],
+  keywords:["deafness", "deaf", "accessibility", "hearing", "impaired", "ear"],
 };
 
 export const dentalFloss: IMXIcon = {
   name: "dental-floss",
-  value: require("./svgs/dental-floss.svg"),
+  value: dentalFlossIcon,
   categories: ["health"],
+  keywords:["dental floss", "floss", "dentist", "tooth", "care", "flossing", "oral", "hygiene", "teeth"],
 };
 
 export const dentalImplant: IMXIcon = {
   name: "dental-implant",
-  value: require("./svgs/dental-implant.svg"),
+  value: dentalImplantIcon,
   categories: ["health"],
+  keywords:["dental", "dentist", "implant", "tooth", "care", "oral"],
 };
 
 export const diabetes: IMXIcon = {
   name: "diabetes",
-  value: require("./svgs/diabetes.svg"),
+  value: diabetesIcon,
   categories: ["health"],
+  keywords:["diabetes", "hand", "finger", "blood", "healtcheck", "endocrinology", "blood test", "blood drop", "blood sugar", "glucose"],
 };
 
 export const dna: IMXIcon = {
   name: "dna",
-  value: require("./svgs/dna.svg"),
+  value: dnaIcon,
   categories: ["health"],
+  keywords:["dna", "genes", "genome", "biology", "genetics"],
 };
 
 export const doctorsBag: IMXIcon = {
   name: "doctors-bag",
-  value: require("./svgs/doctors-bag.svg"),
+  value: doctorsBagIcon,
   categories: ["health"],
+  keywords:["doctor", "bag", "briefcase", "healthcare", "first aid kit", "medical box", "medicine", "emergency", "medical", "hospital", "nurse"],
 };
 
 export const doctor: IMXIcon = {
   name: "doctor",
-  value: require("./svgs/doctor.svg"),
+  value: doctorIcon,
   categories: ["health"],
+  keywords:["doctor", "surgeon", "physician", "md", "medical", "nurse", "gp", "general practitioner"],
 };
 
 export const dose: IMXIcon = {
   name: "dose",
-  value: require("./svgs/dose.svg"),
+  value: doseIcon,
   categories: ["health"],
+  keywords:["dose", "doseage", "dropper", "drop", "tool", "medicine", "sample", "eyedropper"],
 };
 
 export const embryo: IMXIcon = {
   name: "embryo",
-  value: require("./svgs/embryo.svg"),
+  value: embryoIcon,
   categories: ["health"],
+  keywords:["embryo", "baby", "fetus", "motherhood", "pregnancy", "pregnant", "maternity", "mother"],
 };
 
 export const fertilization: IMXIcon = {
   name: "fertilization",
-  value: require("./svgs/fertilization.svg"),
+  value: fertilizationIcon,
   categories: ["health"],
+  keywords:["fertilization", "cell", "insemination", "ovulation", "reproduction", "sperm", "egg"],
 };
 
 export const fitnessWatch: IMXIcon = {
   name: "fitness-watch",
-  value: require("./svgs/fitness-watch.svg"),
+  value: fitnessWatchIcon,
   categories: ["health"],
+  keywords:["fitness", "watch", "smartwatch", "fit band", "heart rate", "wearable", "tracker", "sport", "diet"],
 };
 
 export const gasMask: IMXIcon = {
   name: "gas-mask",
-  value: require("./svgs/gas-mask.svg"),
+  value: gasMaskIcon,
   categories: ["health"],
+  keywords:["gas", "mask", "protection", "safety", "chemical mask", "industrial mask", "respirator"],
 };
 
 export const hearingAid: IMXIcon = {
   name: "hearing-aid",
-  value: require("./svgs/hearing-aid.svg"),
+  value: hearingAidIcon,
   categories: ["health"],
+  keywords:["hearing", "aid", "deaf", "deafness", "ear", "earbuds", "sound"],
 };
 
 export const heartCare: IMXIcon = {
   name: "heart-care",
-  value: require("./svgs/heart-care.svg"),
+  value: heartCareIcon,
   categories: ["health"],
+  keywords:["heart", "care", "check"],
 };
 
 export const heartMonitor: IMXIcon = {
   name: "heart-monitor",
-  value: require("./svgs/heart-monitor.svg"),
+  value: heartMonitorIcon,
   categories: ["health"],
+  keywords:["heart", "pulse", "cardiogram", "beat", "cardiac", "healthcare"],
 };
 
 export const hospitalBed: IMXIcon = {
   name: "hospital-bed",
-  value: require("./svgs/hospital-bed.svg"),
+  value: hospitalBedIcon,
   categories: ["health"],
+  keywords:["hospital", "bed", "care", "medical"],
 };
 
 export const hospitalSign: IMXIcon = {
   name: "hospital-sign",
-  value: require("./svgs/hospital-sign.svg"),
+  value: hospitalSignIcon,
   categories: ["health"],
+  keywords:["hospital", "sign", "medical", "road sign"],
 };
 
 export const hospital: IMXIcon = {
   name: "hospital",
-  value: require("./svgs/hospital.svg"),
+  value: hospitalIcon,
   categories: ["health"],
+  keywords:["hospital", "clinic", "medical", "building", "therapy", "health center"],
 };
 
 export const keepAwayFromChildren: IMXIcon = {
   name: "keep-away-from-children",
-  value: require("./svgs/keep-away-from-children.svg"),
+  value: keepAwayFromChildrenIcon,
   categories: ["health"],
+  keywords:["keep away from children", "keep away children", "not suitable for children", "no children", "keep away from kids", "keep away kids", "not suitable for kids", "no kids", "keep away from babys", "keep away babys", "not suitable for babys", "no babys"],
 };
 
 export const mask: IMXIcon = {
   name: "mask",
-  value: require("./svgs/mask-1.svg"),
+  value: maskIcon,
   categories: ["health"],
+  keywords:["mask", "face mask", "medical mask", "shield mask", "protective mask", "dentist mask", "covid", "prevention"],
 };
 
 export const maskAlt: IMXIcon = {
   name: "mask-alt",
-  value: require("./svgs/mask-2.svg"),
+  value: maskAltIcon,
   categories: ["health"],
+  keywords:["mask", "face mask", "medical mask", "shield mask", "protective mask", "dentist mask", "covid", "prevention"],
 };
 
 export const massage: IMXIcon = {
   name: "massage",
-  value: require("./svgs/massage.svg"),
+  value: messageIcon,
   categories: ["health"],
+  keywords:["massage", "spa", "relax", "body massage", "treatment"],
 };
 
 export const medicalBook: IMXIcon = {
   name: "medical-book",
-  value: require("./svgs/medical-book.svg"),
+  value: medicalBookIcon,
   categories: ["health"],
+  keywords:["medical", "book", "notebook", "healthcare", "cross", "medicine book", "pharmacy book"],
 };
 
 export const medicine: IMXIcon = {
   name: "medicine",
-  value: require("./svgs/medicine.svg"),
+  value: medicineIcon,
   categories: ["health"],
+  keywords:["medicine", "drugs", "medical", "pill"],
 };
 
 export const microorganisms: IMXIcon = {
   name: "microorganisms",
-  value: require("./svgs/microorganisms.svg"),
+  value: microorganismsIcon,
   categories: ["health"],
+  keywords:["microorganisms", "bacteria", "microbes", "virus", "biological", "cell", "biology"],
 };
 
 export const microscope: IMXIcon = {
   name: "microscope",
-  value: require("./svgs/microscope.svg"),
+  value: microscopeIcon,
   categories: ["health"],
+  keywords:["microscope", "lab", "biology", "chemistry", "instrument", "science", "education", "learn"],
 };
 
 export const nurse: IMXIcon = {
   name: "nurse",
-  value: require("./svgs/nurse.svg"),
+  value: nurseIcon,
   categories: ["health"],
+  keywords:["nurse", "healthcare", "medical", "medical assistant", "nursemaid"],
 };
 
 export const pacifier: IMXIcon = {
   name: "pacifier",
-  value: require("./svgs/pacifier.svg"),
+  value: pacifierIcon,
   categories: ["health"],
+  keywords:["pacifier", "baby", "child", "children", "kid", "binky", "dummy", "nipple", "soother", "teether"],
 };
 
 export const pill: IMXIcon = {
   name: "pill",
-  value: require("./svgs/pill.svg"),
+  value: pillIcon,
   categories: ["health"],
+  keywords:["pill", "drugs", "medicine", "capsule", "medical", "remedy", "medicament", "dragee", "treatment", "aid"],
 };
 
 export const plasticSurgery: IMXIcon = {
   name: "plastic-surgery",
-  value: require("./svgs/plastic-surgery.svg"),
+  value: plasticSurgeryIcon,
   categories: ["health"],
+  keywords:["plastic surgery", "cosmetic surgery", "esthetician"],
 };
 
 export const poop: IMXIcon = {
   name: "poop",
-  value: require("./svgs/poop.svg"),
+  value: poopIcon,
   categories: ["health"],
+  keywords:["poop", "shit", "crap", "poo", "faeces"],
 };
 
 export const pregnancyNoAlcohol: IMXIcon = {
   name: "pregnancy-no-alcohol",
-  value: require("./svgs/pregnancy-no-alcohol.svg"),
+  value: pregnancyNoAlcoholIcon,
   categories: ["health"],
+  keywords:["pregnancy", "preagnant", "no alcohol", "no drinking", "non-alcoholic", "do not drink"],
 };
 
 export const pulse: IMXIcon = {
   name: "pulse",
-  value: require("./svgs/pulse.svg"),
+  value: pulseIcon,
   categories: ["health"],
+  keywords:["pulse", "heartbeat", "heartrate"],
 };
 
 export const radioactive: IMXIcon = {
   name: "radioactive",
-  value: require("./svgs/radioactive.svg"),
+  value: radioactiveIcon,
   categories: ["health"],
+  keywords:["radioactive", "biohazard", "danger", "radioactivity", "radiation", "nuke", "atomic bomb", "atom", "nuclear", "energy"],
 };
 
 export const rhNegative: IMXIcon = {
   name: "rh-negative",
-  value: require("./svgs/rh-negative.svg"),
+  value: rhNegativeIcon,
   categories: ["health"],
+  keywords:["blood", "rh", "negative", "minus", "blood group", "blood type", "rhesus", "donorship"],
 };
 
 export const rhPositive: IMXIcon = {
   name: "rh-positive",
-  value: require("./svgs/rh-positive.svg"),
+  value: rhPositiveIcon,
   categories: ["health"],
+  keywords:["blood", "rh", "positive", "plus", "blood group", "blood type", "rhesus", "donorship"],
 };
 
 export const rodOfAsclepius: IMXIcon = {
   name: "rod-of-ascelepius",
-  value: require("./svgs/rod-of-asclepius.svg"),
+  value: rodOfAsclepiusIcon,
   categories: ["health"],
+  keywords:["rod of ascelepius", "staff of ascelepius", "ascelepius"],
 };
 
 export const signLanguage: IMXIcon = {
   name: "sign-language",
-  value: require("./svgs/sign-language.svg"),
+  value: signLanguageIcon,
   categories: ["health"],
+  keywords:["sign", "language", "deaf", "asl", "deafness", "mute", "dumb"],
 };
 
 export const sneeze: IMXIcon = {
   name: "sneeze",
-  value: require("./svgs/sneeze.svg"),
+  value: sneezeIcon,
   categories: ["health"],
+  keywords:["sneeze", "cough", "covid", "corona", "coronavirus", "flu", "infection", "sick", "cold", "hayfever"],
 };
 
 export const stethoscope: IMXIcon = {
   name: "stethoscope",
-  value: require("./svgs/stethoscope.svg"),
+  value: stethoscopeIcon,
   categories: ["health"],
+  keywords:["stethoscope", "line", "doctor", "medical", "nurse", "doctor", "healthcare"],
 };
 
 export const stretcher: IMXIcon = {
-  name: "strecher",
-  value: require("./svgs/stretcher.svg"),
+  name: "stretcher",
+  value: stretcherIcon,
   categories: ["health"],
+  keywords:["stretcher", "aid", "hospital", "patient", "medical", "bed", "cot", "ambulance"],
 };
 
 export const syringe: IMXIcon = {
   name: "syringe",
-  value: require("./svgs/syringe.svg"),
+  value: syringeIcon,
   categories: ["health"],
+  keywords:["syringe", "medicine", "doctor", "nurse", "drugs", "injection", "vaccine", "vaccination", "flu", "medical", "needle", "shot"],
 };
 
 export const teeth: IMXIcon = {
   name: "teeth",
-  value: require("./svgs/teeth.svg"),
+  value: teethIcon,
   categories: ["health"],
+  keywords:["teeth", "false teeth", "tooth", "set of teeth", "dental", "dentist", "gums", "dentistry"],
 };
 
 export const thermometer: IMXIcon = {
   name: "thermometer",
-  value: require("./svgs/thermometer.svg"),
+  value: thermometerIcon,
   categories: ["health"],
+  keywords:["thermometer", "temperature", "weather", "celsius", "fahrenheit", "gauge", "science", "measure", "heat", "cold", "degree"],
 };
 
 export const toiletPaper: IMXIcon = {
   name: "toilet-paper",
-  value: require("./svgs/toilet-paper.svg"),
+  value: toiletPaperIcon,
   categories: ["health"],
+  keywords:["toilet", "paper", "napkins", "roll"],
 };
 
 export const toothCracked: IMXIcon = {
   name: "tooth-cracked",
-  value: require("./svgs/tooth-cracked.svg"),
+  value: toothCrackedIcon,
   categories: ["health"],
+  keywords:["tooth", "cracked tooth", "caries", "dental", "dentist", "broken tooth", "damaged tooth", "teeth"],
 };
 
 export const tooth: IMXIcon = {
   name: "tooth",
-  value: require("./svgs/tooth.svg"),
+  value: toothIcon,
   categories: ["health"],
+  keywords:["tooth", "teeth", "dental", "dentist"],
 };
 
 export const veterinarian: IMXIcon = {
   name: "veterinarian",
-  value: require("./svgs/veterinarian.svg"),
+  value: veterinarianIcon,
   categories: ["health"],
+  keywords:["veterinarian", "vet", "pet", "animal", "clinic", "veterinary"],
 };
 
 export const vial: IMXIcon = {
   name: "vial",
-  value: require("./svgs/vial.svg"),
+  value: vialIcon,
   categories: ["health"],
+  keywords:["vial", "test tube", "lab", "sample", "chemistry"],
 };
 
 export const virusProtection: IMXIcon = {
   name: "virus-protection",
-  value: require("./svgs/virus-protection.svg"),
+  value: virusProtectionIcon,
   categories: ["health"],
+  keywords:["virus", "protection", "shield", "immunity", "corona", "coronavirus", "covid", "bacteria", "safety", "antivirus", "hygiene", "stop virus"],
 };
 
 export const virusStop: IMXIcon = {
   name: "virus-stop",
-  value: require("./svgs/virus-stop.svg"),
+  value: virusStopIcon,
   categories: ["health"],
+  keywords:["virus", "stop virus", "protection", "corona", "coronavirus", "covid", "bacteria", "safety"],
 };
 
 export const virus: IMXIcon = {
   name: "virus",
-  value: require("./svgs/virus.svg"),
+  value: virusIcon,
   categories: ["health"],
+  keywords:["virus", "corona", "coronavirus", "covid", "bacteria", "flu"],
 };
 
 export const visuallyImpaired: IMXIcon = {
   name: "visually-impaired",
-  value: require("./svgs/visually-impaired.svg"),
+  value: visuallyImpairedIcon,
   categories: ["health"],
+  keywords:["visually", "impaired", "blind", "eye", "sight"],
 };
 
 export const weightScale: IMXIcon = {
   name: "weight-scale",
-  value: require("./svgs/weight-scale.svg"),
+  value: weightScaleIcon,
   categories: ["health"],
+  keywords:["weight", "scale", "measure", "bathroom", "bath", "overweight", "obesity", "weighing", "fitness", "weightloss"],
 };
 
 export const health = [
